[FEATURE] Fix id type

+++ enfivy-back/src/main/java/com/envify/back/dao/ConfigPackageFileDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConfigPackageFileDao extends JpaRepository<ConfigPackageFileEntity, Long> {
+public interface ConfigPackageFileDao extends JpaRepository<ConfigPackageFileEntity, Integer> {
 
     ConfigPackageFileEntity getById(int id);
     ConfigPackageFileEntity getByPackageVersionId(int packageVersionId);
